Refactor and split up util.hpp
This reduces rebuild times when touching a utility header
