+++ nucleus/src/data/handle_table.hpp
@@ -10,7 +10,6 @@
 #include <mutex>
 #include <shared_mutex>
 #include <unordered_map>
-#include <util.hpp>
 
 namespace data {
