+++ nucleus/src/data/tracked_object.hpp
@@ -8,9 +8,9 @@
 #include <map>
 #include <memory>
 #include <mutex>
+#include <ref_object.hpp>
 #include <shared_mutex>
 #include <unordered_map>
-#include <util.hpp>
 #include <vector>
 
 namespace data {
