+++ nucleus/src/lifecycle/command_line.cpp
@@ -3,7 +3,6 @@
 #include "argument_iterator.hpp"
 #include "kernel.hpp"
 #include "scope/context_full.hpp"
-#include "util.hpp"
 
 #include <algorithm>
 #include <cstdlib>
