+++ nucleus/src/platform_abstraction/linux/process_manager.cpp
@@ -16,7 +16,6 @@
 #include <thread>
 #include <type_traits>
 #include <unistd.h>
-#include <util.hpp>
 #include <variant>
 
 namespace ipc {
