+++ nucleus/src/scope/context_impl.hpp
@@ -12,7 +12,6 @@
 
 #include <optional>
 #include <shared_mutex>
-#include <util.hpp>
 #include <utility>
 
 namespace config {
