+++ plugin_api/include/buffer_stream.hpp
@@ -1,5 +1,5 @@
 #pragma once
-#include "util.hpp"
+#include "span.hpp"
 
 #include <cstdint>
 #include <ios>
