+++ plugin_api/include/ref_object.hpp
@@ -1,0 +1,81 @@
+#pragma once
+#include <memory>
+namespace util {
+    //
+    // Base class for all "by-reference-only" objects
+    //
+    template<typename T>
+    class RefObject : public std::enable_shared_from_this<T> {
+    public:
+        ~RefObject() noexcept = default;
+        RefObject() noexcept;
+        RefObject(const RefObject &) = delete;
+        RefObject(RefObject &&) noexcept = default;
+        RefObject &operator=(const RefObject &) = delete;
+        RefObject &operator=(RefObject &&) = delete;
+
+        std::shared_ptr<const T> baseRef() const {
+            return std::enable_shared_from_this<T>::shared_from_this();
+        }
+
+        std::shared_ptr<T> baseRef() {
+            return std::enable_shared_from_this<T>::shared_from_this();
+        }
+
+        template<typename S>
+        std::shared_ptr<const S> tryRef() const;
+        template<typename S>
+        std::shared_ptr<const S> ref() const;
+        template<typename S>
+        std::shared_ptr<S> tryRef();
+        template<typename S>
+        std::shared_ptr<S> ref();
+    };
+
+    template<typename T>
+    RefObject<T>::RefObject() noexcept {
+        static_assert(std::is_base_of_v<RefObject, T>);
+    }
+
+    template<typename T>
+    template<typename S>
+    std::shared_ptr<const S> RefObject<T>::tryRef() const {
+        static_assert(std::is_base_of_v<T, S>);
+        if constexpr(std::is_same_v<T, S>) {
+            return baseRef();
+        } else {
+            return std::dynamic_pointer_cast<const S>(baseRef());
+        }
+    }
+
+    template<typename T>
+    template<typename S>
+    std::shared_ptr<S> RefObject<T>::tryRef() {
+        static_assert(std::is_base_of_v<T, S>);
+        if constexpr(std::is_same_v<T, S>) {
+            return baseRef();
+        } else {
+            return std::dynamic_pointer_cast<S>(baseRef());
+        }
+    }
+
+    template<typename T>
+    template<typename S>
+    std::shared_ptr<const S> RefObject<T>::ref() const {
+        std::shared_ptr<const S> ptr{tryRef<S>()};
+        if(!ptr) {
+            throw std::bad_cast();
+        }
+        return ptr;
+    }
+
+    template<typename T>
+    template<typename S>
+    std::shared_ptr<S> RefObject<T>::ref() {
+        std::shared_ptr<S> ptr{tryRef<S>()};
+        if(!ptr) {
+            throw std::bad_cast();
+        }
+        return ptr;
+    }
+} // namespace util
